+++ YBS/APIDefine.cs
@@ -15,13 +15,14 @@ namespace YBS
         public const string GOOGLE_AUTHENTICATION = "api/[controller]" + "/";
         //Member
         public const string MEMBER_CREATE = DEFAULT_ROUTE;
-        public const string MEMBER_DETAIL = DEFAULT_ROUTE + "/{Id}";
+        public const string MEMBER_DETAIL = DEFAULT_ROUTE + "/{id}";
         public const string MEMBER_UPDATE = DEFAULT_ROUTE;
         public const string MEMBER_GET_ALL = DEFAULT_ROUTE;
         //Membership Package
-        public const string MEMBERSHIP_PACKAGE_CREATE = DEFAULT_ROUTE;
-        public const string MEMBERSHIP_PACKAGE_DETAIL = DEFAULT_ROUTE + "/{Id}";
-        public const string MEMBERSHIP_PACKAGE_UPDATE = DEFAULT_ROUTE;
-        public const string MEMBERSHIP_PACKAGE_GET_ALL = DEFAULT_ROUTE;
+        public const string MEMBERSHIP_PACKAGE = "membership-package";
+        public const string MEMBERSHIP_PACKAGE_CREATE = MEMBERSHIP_PACKAGE;
+        public const string MEMBERSHIP_PACKAGE_DETAIL = MEMBERSHIP_PACKAGE + "/{id}";
+        public const string MEMBERSHIP_PACKAGE_UPDATE = MEMBERSHIP_PACKAGE;
+        public const string MEMBERSHIP_PACKAGE_GET_ALL = MEMBERSHIP_PACKAGE;
     }
 }
