+++ YBS/Controllers/AccountsController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using YBS.Authorization;
 using YBS.Data.Enums;
+using YBS.Middlewares;
 using YBS.Service.Services;
 using YBS.Services.Dtos.PageRequests;
