+++ YBS/Controllers/AuthenticationController.cs
@@ -1,7 +1,6 @@
 using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using YBS.Authorization;
 using YBS.Data.Enums;
 using YBS.Data.Models;
 using YBS.Service.Services;
