+++ YBS/Middlewares/RoleAuthorizationAttribute.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 
-namespace YBS.Authorization
+namespace YBS.Middlewares
 {
     public class RoleAuthorizationAttribute : TypeFilterAttribute
     {
